+++ app/static/js/session.js
@@ -29,6 +29,9 @@ function updateSessionId(newId) {
   if (typeof loadRecentValues === 'function') {
     loadRecentValues().catch(() => {});
   }
+  if (typeof loadScheduleAutocompleteHints === 'function') {
+    loadScheduleAutocompleteHints().catch(() => {});
+  }
   if (typeof refreshWorkspaceFileCache === 'function') {
     refreshWorkspaceFileCache().catch(() => {});
   }
@@ -55,6 +58,7 @@ window.addEventListener('storage', (e) => {
     if (typeof loadSessionPreferences === 'function') loadSessionPreferences().catch(() => {});
     if (typeof loadSessionVariables === 'function') loadSessionVariables().catch(() => {});
     if (typeof loadRecentValues === 'function') loadRecentValues().catch(() => {});
+    if (typeof loadScheduleAutocompleteHints === 'function') loadScheduleAutocompleteHints().catch(() => {});
     if (typeof refreshWorkspaceFileCache === 'function') refreshWorkspaceFileCache().catch(() => {});
     if (typeof window.refreshActiveProjectContext === 'function') window.refreshActiveProjectContext().catch(() => {});
     if (typeof _updateOptionsSessionTokenStatus === 'function') _updateOptionsSessionTokenStatus();